Avoid `#steps` boilerplate by prepending on `#call`
We get rid of the necessity to wrap the operations' flow with the
`#steps`'s block by prepending a module that decorates the `#call`
method.

If before we had to do:

```ruby
class CreateUser < Dry::Operation
  def call(input)
    steps do
      attributes = step validate(input)
      step create(attributes)
    end
  end

  # ...
end
```

Now we can do:

```ruby
class CreateUser < Dry::Operation
  def call(input)
    attributes = step validate(input)
    step create(attributes)
  end

  # ...
end
```

We want to provide that as the default behavior to improve the
ergonomics of the library. However, we also want to provide a way to
customize or opt out of this magic behavior. Users can inherit from,
e.g., `Dry::Operation[prepend: :run]` to decorate the `#run` method
(multiple methods are allowed by passing an array), or
`Dry::Operation[prepend: false]` to disable the behavior:

```ruby
class CreateUser < Dry::Operation[prepend: :run]
  def run(input)
    attributes = step validate(input)
    step create(attributes)
  end

  # ...
end
```

```ruby
class CreateUser < Dry::Operation[prepend: :false]
  def call(input)
    steps do
      attributes = step validate(input)
      step create(attributes)
    end
  end

  # ...
end
```

Notice that as a consequence of this change, we're "polluting" the
ancestry chain of the descendant classes with the prepender logic:

```ruby
Class.new(Dry::Operation).ancestors
 # [#<Module:0x00007f8e523c1698>, <- Actual prepender
 #  #<Class:0x00007f8e523c1918>, <- Descendant
 #  #<Dry::Operation::Prepender:0x00007f8e523c17d8>, <- Stateful module keeping the prepender options
 #  Dry::Operation,
 #  Dry::Operation::Mixin,
 # ...]
```

When creating a descendant from a descendant, the prepender logic is
duplicated to override the decorated method from the nested descendant
(which goes before the first prepender in the ancestry):

```ruby
Class.new(Class.new(Dry::Operation)).ancestors
 # [#<Module:0x00007f8e52331408>, <- Actual prepender
 #  #<Class:0x00007f8e52331688>, <- Nested descendant
 #  #<Dry::Operation::Prepender:0x00007f8e52331548>, <- Stateful module keeping the prepender options
 #  #<Module:0x00007f8e52331868>, <- Previous prepender
 #  #<Class:0x00007f8e52331ae8>, <- Descendant
 #  #<Dry::Operation::Prepender:0x00007f8e523319a8>, <- Previous stateful module keeping the prepender options
 #  Dry::Operation,
 #  Dry::Operation::Mixin,
 #  ...]
```

Also, notice that `Dry::Operation[]` will return an anonymous `Class` so
it can be inherited from, so `Dry::Operation` won't appear in the
ancestry chain:

```ruby
Class.new(Dry::Operation[prepend: false]).ancestors
 # [#<Class:0x00007f8e5233eec8>, <- Descendant
 #  #<Class:0x00007f8e5233f328>, <- Anonymous class returned by Dry::Operation[]
 #  Dry::Operation::Mixin,
 #  ...]
```

To accomodate the above, the steps logic has been moved to a
`Dry::Operation::Mixin` module so it can be reused both by descendants
from `Dry::Operation` and `Dry::Operation[]`.

It's worth it to highlight all that complexity, although it should be
transparent to the user. As said, it's a trade-off we consider worth it
to improve developer experience.