feat: Add extension for ActiveRecord transactions (#18)
* feat: Add extention for ActiveRecord transactions

This commit introduces a new extension to Dry::Operation that allows
operations to be wrapped in ActiveRecord transactions. This is useful
for ensuring atomicity of operations that involve multiple steps that
need to be rolled back in case of failure. The extension provides a
`transaction` method that can be used to wrap steps in a transaction.
The transaction will be rolled back if any of the steps return a
`Dry::Monads::Result::Failure`.

The extension also supports specifying a custom ActiveRecord class to
initiate the transaction, which is useful when working with multiple
databases.

It also supports additional options for the ActiveRecord transaction.
These options can be set as default for all transactions or overridden
at runtime.