Define #transaction in their extension modules
Rather than defining #transaction on the operation class itself, define it in the extension modules that are being included.

This allows the user to be able to define their own transaction and call `super` to get to the method provided by the module, which is the standard behaviour for modules and Ruby method lookups.