Define transaction methods in extension modules only (#33)
Rather than the extension modules defining `#transaction` on the operation class itself, define it within the modules themselves, which still makes it available on the operation class.

Doing this allows the user to be able to define their _own_ `#transaction` on their operation classes, and then call `super` to get to the method provided by the module, which is the standard behaviour for modules and Ruby method lookups.